scripts/slack-bot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
anthropic>=0.40.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
schedule>=1.2.0
